- precise selection of edges - kill without error the current dot process when startting a new one - rename some methods to avoid lexical hiding warnings - new message boxes and new whatsthis that were delayed during 4.0 string freeze - individual attribute changing mechanism - new signals to inform of the current selection - compiler warnings removal

+++ src/part/DotGraphParsingHelper.cpp
@@ -41,7 +41,7 @@
 
 using namespace std;
 
-#define KGV_MAX_ITEMS_TO_LOAD std::numeric_limits<size_t>::max()
+#define KGV_MAX_ITEMS_TO_LOAD std::numeric_limits<int>::max()
 
 extern DotGraphParsingHelper* phelper;
